+++ empulse/models/proflogit.py
@@ -1,11 +1,8 @@
-from typing import Any, ClassVar
+from typing import ClassVar
 
-from scipy.optimize import OptimizeResult
+from empulse.optimizers import GeneticAlgorithmOptimizer, Optimizer
 
-from empulse.optimizers import GeneticAlgorithmOptimizer
-
-from .._types import FloatNDArray
-from ..metrics import LogitObjective, MaxProfit
+from ..metrics import MaxProfit
 from ._base import BaseLogitClassifier
 from .csclassifier import MetricStrategyFactory
 
@@ -128,8 +125,4 @@ class ProfLogitClassifier(BaseLogitClassifier):
     """
 
     _default_metric_strategy: ClassVar[MetricStrategyFactory] = MaxProfit
-
-    def _optimize(self, objective: LogitObjective, X: FloatNDArray, **kwargs: Any) -> OptimizeResult:
-        """Optimize the objective function using the Real-coded Genetic Algorithm."""
-        optimize = GeneticAlgorithmOptimizer() if self.optimizer is None else self.optimizer
-        return optimize(objective=objective, X=X, **kwargs)
+    _default_optimizer: ClassVar[type[Optimizer]] = GeneticAlgorithmOptimizer